flask_disease_detection/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
Werkzeug==2.3.6